use masks
